chore(client): restructure crate
