+++ cirrus-lib/crates/client/src/audio_player/mod.rs
@@ -1,0 +1,5 @@
+mod player;
+mod sample;
+pub mod state;
+
+pub use player::AudioPlayer;
