+++ cirrus-lib/crates/client/src/lib.rs
@@ -1,6 +1,5 @@
-pub mod audio;
+pub mod audio_player;
 pub mod request;
-pub mod state;
-// pub mod data;
+mod dto;
 
-// pub use cirrus_grpc;
+pub use audio_player::AudioPlayer;
